requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
Werkzeug==2.3.7